feat/2: refactor - use constants for error and success messages, move user methods that interact with database to user repository
