+++ api/src/messages.ts
@@ -1,0 +1,13 @@
+export const ERROR_MESSAGE = {
+	AUTH_BEARER_TOKEN_NOT_EXIST: 'must have a bearer token',
+	AUTH_TOKEN_INVALID_FORMAT: 'invalid bearer token format',
+	AUTH_TOKEN_INVALID: 'token is invalid',
+	AUTH_INVALID_TOKEN_PAYLOAD: 'type of JWT payload is invalid',
+	LOGIN_INVALID_CREDENTIAL: 'email or password is not valid',
+	SERVER_ERROR: 'something unexpected happened on the server',
+};
+
+export const SUCCESS_MESSAGE = {
+	USER_CREATE_SUCCESS: 'user created successfully',
+	LOGIN_SUCCESS: 'login success',
+};
