+++ api/src/repositories/index.ts
@@ -1,0 +1,1 @@
+export * from './user.repository.ts';
